Just to give you an idea of what I'm working on, the "ticker class".
This class will contain a timer which counts up to 500ms, at which point
it executes whatever game actions (functions) are queued to enact. e.g.
Pressing up during a countup triggers the player to move upwhen the
timer reaches 500ms. It then resets.

Its name "Ticker" is a reference to game ticks, a time slot where a
calculation is performed.